Auto report only in motion states, on state change and modal change
